EmailController 기능 분리
