--- MemoryMe/src/main/java/memme/memoryme/auth/api/dto/email/PwChangeRequestDTO.java
@@ -1,11 +1,0 @@
-package memme.memoryme.auth.api.dto.email;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class PwChangeRequestDTO {
-    private String currentPassword;
-    private String newPassword;
-}
