--- MemoryMe/src/main/java/memme/memoryme/auth/api/dto/email/PwChangeResponseDTO.java
@@ -1,13 +1,0 @@
-package memme.memoryme.auth.api.dto.email;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class PwChangeResponseDTO {
-    private String email;
-    private String message;
-}
