+++ MemoryMe/src/main/java/memme/memoryme/auth/api/dto/email/RegisterRequestDTO.java
@@ -8,4 +8,5 @@ import lombok.Setter;
 public class RegisterRequestDTO {
     private String email;
     private String password;
+    private String userName;
 }
