+++ MemoryMe/src/main/java/memme/memoryme/auth/api/dto/email/RegisterResponseDTO.java
@@ -10,4 +10,5 @@ import lombok.Setter;
 public class RegisterResponseDTO {
     private String message;
     private String email;
+    private String userName;
 }
